refactor rename FormComonent.tsx to FormComponent.tsx
