+++ packages/dm-core-plugins/src/index.tsx
@@ -5,7 +5,7 @@ import { TabsContainer } from './tabs/TabsContainer'
 import { MermaidComponent } from './mermaid-plugin/MermaidComponent'
 import { JobInputEdit } from './job/JobInputEdit'
 import { JobControlWrapper } from './job/JobControlWrapper'
-import { FormComponent } from './form/FormComonent'
+import { FormComponent } from './form/FormComponent'
 import Header from './header/Header'
 import { DefaultPreviewComponent } from './default-preview/DefaultPreviewComponent'
 import { DefaultPdfComponent } from './default-pdf/DefaultPdfComponent'
